Extracted generic client makeApp function
